feat(chaining-types): add primitive and complex types in chaining (#6536)
